Updated page count warning and a check for empty filter lists
This commit includes updates to all of the functions to issue a
warning if the maximum number of requests has been reached.
Additionally a check has been added to most of the functions to look
for an empty filter list and return just the first page of results if
no filters are provided.  This should prevent fetching a large number
of rows from the server if the filter list is accidentally omitted or
malformed.