Added functions to show and check filters against api endpoints
The filters available on api endpoints are available through a swagger
api. This commit includes functions that will display the available
endpoints and issue a warning if a filter is provided that is not
known to exist. The swagger endpoint only works for FN_portal filters,
but other apps could be added as they become available.  This commit
also includes a large number of small changes created by the styler
package. The packages passes the check without issue.