Added prepare_payload utility function
This commit extract the code to remove id and slug into a separate
helper function prepare_payload(). prepare_load also transforms the
names of the fields in the returned data-frame to uppercase so that
they match FN-2 fields as closely as possible.  There are no errors or
warnings issued by check() and the package builds successfully.