Added functions to fetch data from creel portal api
This is a big commit and incldues a number of files to fetch data from
the creel portal.  They are very similar to the fn_portl counterparts,
but their associated documentation has been updated to reflect the
differences between the two data sources. This commit also inlcudes
updates to some of the utility functions to accomodate the new api
data or schema endpoints.