Delete get_PTAbstract.R
As noted yesterday, we can re-merge this function later if we want. It might be a better fit in your reporter package.  There are several other functions to get (data) and fetch (download) milestone reports and associated files from project tracker  that I will be creating in the very near future. We will have to give some though to where they belong as well.  It would probably make sense to keep put them in the same package as as get_PtAbstacts.