Updated show_filters(), package contributors, and build errors
This commit updates the show filters function to report all of the
available api endpoints if no endpoint is provided. It also updates
the names of package contributors to include both Jeremy and
Rachel. Finally a number of small updates are included that were
identified by the check() command.  Package successfully build and
tagged with version v2022.09.02.