trycatch in get_credentials, style_pkg() changes
This commit wraps the get_credentials function in a try catch - the
Rstudioapi is used by default, but readlines is now used as a fallback
if rstudioapi is not in use or available.  This commit also includes
changes automatically performed by styler::style_pkg().  Virtually all
of the changes made by styler were the removal of whitespace and the
addition of line breaks.