updated documentation to use \code{} and \url{}
This commit makes a some small changes to the documentation associated
with a large number of files.  Every instance of ~show_filters(...)~
was replaced with \code{show_filters(...)} and every reference the
glis swagger docs were wrapped in \url{}. All of the checks pass and
the packages was successfully compiled to a binary library.