Fixed spurious authentication error
This commit fixes the authentication error that was being reported to
most users. It was being caused by a local assignment to a variable
'token' as the return value of get_token() (which is always null). The
local variable was masking the global value, triggering the warning
message. Everything now works as expected and the tests all pass.