Added optional verbose_sqlsave argument to populate template
This commit adds an optional verbose_sqlsave argument to the populate
template function. This shouldn't be needed in most cases, but has
proven to be valuable when there are issues inserting data into the
target database.  When set to TRUE the actual sql statements and the
databases response are printed to the console.  All tests continue to
pass.