Updated api_to_dataframe to report the number of records
This commit updates the api_to_dataframe function to report the number
of records that will be returned.  This should provide users with an
opportunity to identify when a malformed filter has been submitted and
give them a chance to stop the request before it iterates unnecessarily.
The package pass all of the checks.