Added functionality to insert Fn121_GPS_tracks into template db
this commit adds functionality to add FN121_GPS_Tracks into the
template database.  This was originally missed in the populate
template function, but was complicated because of the field named
'TIMESTAMP' which is a sql keyword that is not escaped properly by
the RODBC package.  The package version was also updated to today's
date. All of the tests pass, and the package passes all of the checks
and build properly.