Version 2025-02-26
This is a bug fix commit. There was an bug in the populate_template
that was stopping the process if the project did not have any
FN121_GPS_Tracks (which is most projects).  This commit also includes
a welcome message that prints out the package version to help us keep
our users up to date.  All tests and checks pass, and the package can
be build successfully.